Moved search in modal
